Add files via upload
This project uses CDN Links for bootstrap and Font awsome, so you can make use of these links. A CDN (Content Delivery Network) is a highly-distributed platform of servers that helps minimize delays in loading web page content by reducing the physical distance between the server and the user. This helps users around the world view the same high-quality content without slow loading times. In this project internal stylesheet is used for styling purpose. An internal stylesheet holds CSS rules for the page in the head section as well as outside but before body tag of the HTML file. The rules only apply to that page, but you can configure CSS classes and IDs that can be used to style multiple elements in the page code. Again, a single change to the CSS rule will apply to all tagged elements on the page.